Fonts, keyboard raw mode, display helpers
Fonts:

- Drop legacy bitmap fonts: FreeMono5pt7b.h and FreeSans7pt7b.h
  removed (both replaced in the widget layer by *_aa fonts for UI
  text and by the Spleen family for game HUDs).
- New Spleen{6x12,8x16,12x24,16x32}.h: monospace bitmaps used by
  the FontSize::TINY/SMALL/MEDIUM/LARGE slots for games and the
  text capture path.
- InterAA9.h: new 9-pixel Inter AA variant that backs the
  `tiny_aa` font slot used by map labels, chat meta lines, and
  hint rows in settings. InterAA11/13/17 regenerated with the
  updated tools/gen_aa_font.py so ascent + advance widths match
  Inter's metrics more accurately (fixes "Desktop" grouping
  "Des"|"kto"|"p" at 17 px).
- src/hardware/aa_font.{cpp,h}: Q8.8 subpixel pen across a whole
  string so fractional advances accumulate cleanly instead of
  rounding per glyph.

Keyboard raw matrix mode:

- src/hardware/keyboard.{cpp,h}: add "raw" mode alongside "normal".
  Raw mode bypasses the default character decoder and surfaces the
  full 12-byte matrix (keyboard_matrix.h), exposing the state of
  every key every frame. Used by the screen editor, the chord
  detector, and the matrix test screen.
- Interrupt mode retained and now dispatches events off the GPIO
  falling edge instead of polling.
- src/lua/bindings/display_bindings.cpp: misc helpers that pair
  with the screen/dialog modules landing alongside this commit.

Font tooling:

- tools/gen_bdf_font.py: parse an X11 BDF font and emit a packed
  Spleen*-style header.
- tools/gen_mono_bitmap.py: convert a PNG glyph strip into the
  same header format for custom monospace faces.
- tools/gen_aa_font.py: tighter Q8.8 advance emission, proper
  ascent handling.

+++ src/hardware/aa_font.h
@@ -17,7 +17,7 @@ struct Glyph {
     uint8_t  h;
     int8_t   xo;
     int8_t   yo;
-    uint8_t  adv;
+    uint16_t adv;   // Q8.8 pixels — see gen_aa_font.py
     uint16_t off;
 } __attribute__((packed));
